picozi/requirements.txt
Pillow == 7.0.0
altair == 4.1.0
matplotlib == 3.3.3
numpy == 1.19.5
opencv_python--headless
scikit_learn == 0.24.1
streamlit == 0.75.0
